Functioning sensor, LED, and Lux 7-Seg
Tsing setpoint/PID struct updating in progress.

+++ src/pidtask.h
@@ -18,12 +18,12 @@ typedef struct {
     float Kp;           // proportional gain
     float Ki;           // integral gain
     float Kd;           // derivative gain
-    float setpoint;     // intensity setpoint
+    uint16_t setpoint;  // intensity setpoint
     float integral;     // error acccumulator, sum of errors over time
-    float prev_error;   // previous error, used for (d_error/dt)
+    uint16_t prev_error;   // previous error, used for (d_error/dt)
     float delta_t;      // change in time for derivative
-    float max_lim;      // value to set upper rail for output
-    float min_lim;      // value to set lower rail for output
+    uint16_t max_lim;      // value to set upper rail for output
+    uint8_t min_lim;      // value to set lower rail for output
 }PID_t;
 
 // prototype for function that initializes PID structure
